BracketCityPuzzle

Documentation for BracketCityPuzzle.

These are some software tools for working with "Bracket City" puzzles as appear onnthe web site for The Atlantic Monthly.

Example

using BracketCityPuzzle

# Puzzle from 2025-09-07:
puzzle = """ "Gari[like Larry [underdog vs. Goliath] or Ma[wearing one [elp[ out your credit card (bad idea, usually)]e rof ,eulc siht] is a good way to connect with the [[fancy kind of pen 🖋️] of  (👵 ➡️ 👧)]]ma Gandhi 🦅]i enters [time resisted by a t[weird or not di[able to be [\"As  on TV\" 📺]] by [customary number of weeks of [see and re[to \"get the \" of something is to under[\"I  [word that will appear if you get this clue right]ed\"] the basic idea]er] given before leaving a job]]ler, maybe 💤]les" """

parsed = parse_puzzle(puzzle)

show_puzzle(parsed)

 "Gari
[
  like Larry
  [
    underdog vs. Goliath
  ]
   or Ma
  [
    wearing one
    [
      elp
      [
         out your credit card (bad idea, usually)
      ]
      e rof ,eulc siht
    ]
     is a good way to connect with the
    [
      [
        fancy kind of pen 🖋️
      ]
       of  (👵 ➡️ 👧)
    ]
  ]
  ma Gandhi 🦅
]
i enters
[
  time resisted by a t
  [
    weird or not di
    [
      able to be
      [
        "As  on TV" 📺
      ]
    ]
     by
    [
      customary number of weeks of
      [
        see and re
        [
          to "get the " of something is to under
          [
            "I
            [
              word that will appear if you get this clue right
            ]
            ed"
          ]
           the basic idea
        ]
        er
      ]
       given before leaving a job
    ]
  ]
  ler, maybe 💤
]

Add some answers:

only(findBracket("kind of pen", parsed)).answer = "fountain"

only(findBracket("Goliath", parsed)).answer = "David"

only(findBracket("on TV", parsed)).answer = "seen"

only(findBracket("able to", parsed)).answer = "visible"

only(findBracket("out your", parsed)).answer = "max"

only(findBracket("👧", parsed)).answer = "youth"

only(findBracket(",eulc", parsed)).answer = "backwards"

only(findBracket("wearing", parsed)).answer = "hat"

only(findBracket("Larry", parsed)).answer = "bald"

only(findBracket("clue right", parsed)).answer = "corrected"

show_puzzle(preduce(parsed))

 "Garibaldi enters
[
  time resisted by a t
  [
    weird or not divisible by
    [
      customary number of weeks of
      [
        see and re
        [
          to "get the " of something is to under
          [
            "I  correcteded"
          ]
           the basic idea
        ]
        er
      ]
       given before leaving a job
    ]
  ]
  ler, maybe 💤
]

These are the remaining answers, but we don't het have a way to identify the Brackets to associate them with: "stand", "gist", "notice", "two", "odd", "nap"

BracketCityPuzzle.Bracket

BracketCityPuzzle.findBracket

BracketCityPuzzle.parse_puzzle

BracketCityPuzzle.preduce

BracketCityPuzzle.show_puzzle

BracketCityPuzzle.to_string

BracketCityPuzzle.tokenize_puzzle

BracketCityPuzzle.Bracket — Type

Bracket(clue, answer)

Represents a bracketed clue, and possibly its answer.

BracketCityPuzzle.findBracket — Function

findBracket(test, ::BCPuzzle)::Vector{Bracket}

Returns a vector of Bracket objects that match test, which can be a predicate or a substring to match.

BracketCityPuzzle.parse_puzzle — Function

parse_puzzle(string_or_token_vector)

Parses the string representation of a Bracket City puzzle. The result is a vector of strings and nested Bracket objects.

BracketCityPuzzle.preduce — Function

preduce(::BCPuzzle)::BCPuzzle

Simplifies the puzzle to a new puzzle by substituting in any Bracket answers that are known.

BracketCityPuzzle.show_puzzle — Method

show_puzzle(parsed::BCPuzzle)

Outputs the puzzle in a more readable format with lines indented proportional to bracket nesting.

If a Bracket has an answer then it is printed immediately before the close braclet.

BracketCityPuzzle.to_string — Function

to_string(puzzle::BCPuzzle)::AbstractString

Convert the puzzle to its textual representation. This is the inverse of parse_puzzle.

BracketCityPuzzle.tokenize_puzzle — Method

tokenize_puzzle(::AbstractString)

Tokenizes the string to produce a Vector of tokens consisting of strings, :openbracket and :closebracket symbols.